Split test_cloud_record_summary.py into two files
- tests in api/tests/test_cloud_record_summary_get.py make get requests and looks
  at the response for test pass/fail
- api/tests/test_cloud_record_summary_helper.py test helper methods and looks at
  the result of those method to determine test pass/fail (note, some of these do
  make get request objects, as the helper methods expect take the whole request as
  input)
